Implement Installed Themes Checker based on Cache
Why did I decide to do this? The application launches and checks cache
for available “installed” themes. It is never recommended to run the
application on an empty cache, especially if you are running the color
picker.

For now let’s use a toast to warn the user, but as a fail case, this
should completely disable the application from user interaction except
to reboot.

We could use Reflection or CMSDK but that won’t be necessary. I will
develop on this more-simple code in the future.